+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix issue with asset path ordering. See [#96](https://github.com/dockyard/ember-appkit-rails/issues/96) for details.
 * Fix component generator to not add \_component suffix
+* Bootstrap generator remove jbuilder
 
 ## 0.2.0